Module skema::database

Interface to the graph database we are using for persisting GroMEt objects and performing queries on them. We currently use MemgraphDB, an in-memory graph database. ! Currently the literals in the opening main function are getting double wired into the + primitive, guessing some external wiring bug…

Structs

ConstructorArgs

Edge

MetadataNode

Node

Functions

create_att_abstract

create_att_expression

create_att_literal

create_att_predicate

create_att_primitive

create_conditional

create_for_loop

create_function

create_import

create_opi

create_opo

create_while_loop

cross_att_wiring

execute_query

external_wiring

import_wiring

internal_wiring

parse_gromet_queries

wff_cross_att_wiring

wff_wiring

wfopi_cross_att_wiring

wfopi_wiring

wfopo_cross_att_wiring

wfopo_wiring

wopio_wiring