Module skema::model_extraction

Structs

ACSet

ModelEdge

This struct is the edge struct for the constructed petgraph

ModelNode

This struct is the node struct for the constructed petgraph

Term

A product of rate and species that is added or subtracted. THere should just be one rate, but since we’re parsing and there could be noise, this accommodates possibly reading several names of things that should be combined into a single rate.

Functions

find_pn_dynamics

This function finds the core dynamics and returns a vector of node id’s that meet the criteria for identification

get_args_MET

This is a recursive function that walks along the wired subgraph of an expression to construct the expression tree

get_operator_MET

This gets the operator from the node name

get_subgraph

This function takes in a node id and returns the nodes and edges in it

module_id2mathml_MET_ast

This is the main function call for model extraction.

subgrapg2_core_dyn_MET_ast

Once the function node has been identified, this function takes it from there to extract the vector of FirstOrderODE’s

to_primitive

to_terminal

trim_calls