Function skema::model_extraction::find_pn_dynamics

pub async fn find_pn_dynamics(module_id: i64, config: Config) -> Vec<i64>

This function finds the core dynamics and returns a vector of node id’s that meet the criteria for identification

Based on the fact we are getting in only the function we expect to have dynamics, this should just be depricated in the future and replaced with a inference to move from the module_id to the top level function node id, however for now we should keep it as a simple heuristic because it is used in the original code2amr (zip repo) endpoint which would need to be updated first.

Plus the case when it fails defaults to a emptry AMR which is preferable to crashing.